hearing-prep/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "hearing-prep"
version = "0.1.0"
description = "Hearing Prep - Oral Q&A simulation for ICE cross-examination practice"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.41.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "google-cloud-speech>=2.21.0",
    "boxsdk>=3.12.0",
    "simple-salesforce>=1.12.0",
    "anthropic>=0.42.0",
    "fpdf2>=2.8.0",
]
